Moved password components to separate folder
